clean up pytensor.graph.sched
